Create Poc.vb
An attacker can access all files, images and etc. without the passcode.  the passcode can be easily bypassed by  triggering  an unhandled situation in the GUI
The vb.net app provided is used to exploit the vulnerability.
to exploit:
1. Compile and run the app
2. Close Soroush IM messgaer window via X button (NOT VIA SYSTEM TRAY OR FORCE CLOSE)
3. Reopen the app
4. Passcode bypassed
Compile via.NET4 and run on the victim system running windows (others?)
NOTE: be sure that the Soroush messager window is not minimized.